feat: migrated all of the code from the old ExampleProxySQLSetup repo
**Description**

This code has diverged enough from the orignial repo that it's no longer
really a fork. I've copied the 1.0 tag from ExampleProxySQLSetup into
this new repo and will tag this as 1.0 as well.

The 1.0 tag will be using the ruby scripts we have that mangage the
cluster state, while future versions will start using the proxysql-agent.